requirements.txt
# Lighter量化交易程序依赖包

# 核心依赖
lighter>=1.0.0
asyncio
aiohttp>=3.8.0
websockets>=10.0

# 以太坊账户管理
eth-account>=0.13.0

# 数据验证
pydantic>=2.0.0

# 数据处理
numpy>=1.21.0
pandas>=1.3.0

# 数学计算
scipy>=1.7.0

# 配置管理
PyYAML>=6.0

# 日志
colorlog>=6.0.0

# 通知功能
# smtplib 是Python内置模块，无需额外安装

# GUI界面依赖
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0
PySide6>=6.4.0

# 图表库
mplfinance>=0.12.0
plotly>=5.0.0
pyqtgraph>=0.13.0

# 技术指标
ta-lib>=0.4.0

# GUI主题和图标
qdarkstyle>=3.0.0
qtawesome>=1.2.0

# 网络通信
requests>=2.28.0

# 可视化（可选）
matplotlib>=3.5.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0

# 开发工具
black>=22.0.0
flake8>=4.0.0
mypy>=0.950